+++ Examples/Basic iOS Demo/iCarouselExampleViewController.m
@@ -110,7 +110,7 @@
                                                        delegate:self
                                               cancelButtonTitle:nil
                                          destructiveButtonTitle:nil
-                                              otherButtonTitles:@"Linear", @"Rotary", @"Inverted Rotary", @"Cylinder", @"Inverted Cylinder", @"Wheel", @"Inverted Wheel", @"CoverFlow", @"CoverFlow2", @"Time Machine", @"Custom", nil];
+                                              otherButtonTitles:@"Linear", @"Rotary", @"Inverted Rotary", @"Cylinder", @"Inverted Cylinder", @"Wheel", @"Inverted Wheel", @"CoverFlow", @"CoverFlow2", @"Time Machine", @"Inverted Time Machine", @"Custom", nil];
     [sheet showInView:self.view];
     [sheet release];
 }
@@ -155,16 +155,19 @@
 
 - (void)actionSheet:(UIActionSheet *)actionSheet didDismissWithButtonIndex:(NSInteger)buttonIndex
 {
-    //map button index to carousel type
-    iCarouselType type = buttonIndex;
-    
-    //carousel can smoothly animate between types
-    [UIView beginAnimations:nil context:nil];
-    carousel.type = type;
-    [UIView commitAnimations];
-    
-    //update title
-    navItem.title = [actionSheet buttonTitleAtIndex:buttonIndex];
+    if (buttonIndex >= 0)
+    {
+        //map button index to carousel type
+        iCarouselType type = buttonIndex;
+        
+        //carousel can smoothly animate between types
+        [UIView beginAnimations:nil context:nil];
+        carousel.type = type;
+        [UIView commitAnimations];
+        
+        //update title
+        navItem.title = [actionSheet buttonTitleAtIndex:buttonIndex];
+    }
 }
 
 #pragma mark -